Stupid idea :/
I edited some of the code for the cvlib package to make certain features work in my Object Detection program. (Therefore the current python file will not work if compiled and will only work in VSCode or similar editor with this package in place of the object_detection.py included in cvlib). This will be rectified ASAP